pyproject.toml
[project]
name = "travel-agent-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.20.0",
    "litellm>=1.80.5",
]
